W2A-95: Move usePaddle to hooks/use-paddle
